Refactor into base and image classes
Base class does not assume a full image exists, includes only capability
for generating header/footer.
